requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Whisper & ML
openai-whisper
torch
torchaudio
numpy
tiktoken
more-itertools

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Configuration
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# File handling
aiofiles==23.2.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Utilities
python-dateutil==2.8.2
